CHE-58: Refactoring. Introducing new classes for handling Kubernetes resources (KubernetesContainer, KubernetesService, KubernetesEnvVar)
